Migrated sample app to use a navigation drawer to properly show tokens and requests
