+++ twilio-auth-sample/src/main/AndroidManifest.xml
@@ -10,32 +10,28 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".main.MainActivity"
-            android:screenOrientation="portrait"></activity>
-
-        <activity
-            android:name=".approvalrequests.ApprovalRequestsListActivity"
-            android:label="@string/title_activity_approval_requests_list"
-            android:screenOrientation="portrait"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity
             android:name=".registration.RegistrationActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
+            android:name=".main.MainActivity"
+            android:label="@string/menu_navigation_requests"
+            android:theme="@style/AppTheme.NoActionBar" />
+        <activity
             android:name=".approvalrequests.detail.ApprovalRequestDetailActivity"
             android:label="@string/title_activity_approval_request_detail"
-            android:parentActivityName=".approvalrequests.ApprovalRequestsListActivity"
+            android:parentActivityName=".main.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.twilio.authsample.approvalrequests.ApprovalRequestsListActivity" />
+                android:value="com.twilio.authsample.main.MainActivity" />
         </activity>
 
         <!-- Push notifications -->
