+++ twilio-auth-sample/src/main/java/com/twilio/authsample/approvalrequests/events/ApprovalRequestsUpdatedEvent.java
@@ -3,7 +3,7 @@ package com.twilio.authsample.approvalrequests.events;
 /**
  * This class represents a change on the ApprovalRequests or on the
  * filters used by each fragment.
- * The actual changes are stored by the ApprovalRequestsListActivity, this just notifies subscribers to
+ * The actual changes are stored by the RequestsFragment, this just notifies subscribers to
  * fetch a new value
  * <p/>
  * Created by jsuarez on 4/22/16.
