--- twilio-auth-sample/src/main/res/layout/activity_approval_requests_list.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                                 xmlns:app="http://schemas.android.com/apk/res-auto"
-                                                 xmlns:tools="http://schemas.android.com/tools"
-                                                 android:id="@+id/main_content"
-                                                 android:layout_width="match_parent"
-                                                 android:layout_height="match_parent"
-                                                 android:background="@color/background_color"
-                                                 android:fitsSystemWindows="true"
-                                                 tools:context=".approvalrequests.ApprovalRequestsListActivity">
-
-    <android.support.design.widget.AppBarLayout
-            android:id="@+id/appbar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingTop="@dimen/appbar_padding_top"
-            android:theme="@style/AppTheme.AppBarOverlay">
-
-        <android.support.v7.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:background="?attr/colorPrimary"
-                app:layout_scrollFlags="scroll|enterAlways"
-                app:popupTheme="@style/AppTheme.PopupOverlay">
-
-        </android.support.v7.widget.Toolbar>
-
-        <android.support.design.widget.TabLayout
-                android:id="@+id/tabs"
-                style="@style/TabLayoutStyle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
-
-    </android.support.design.widget.AppBarLayout>
-
-    <LinearLayout android:layout_width="match_parent"
-                  android:layout_height="match_parent"
-                  android:orientation="vertical"
-                  app:layout_behavior="@string/appbar_scrolling_view_behavior">
-        <TextView
-                android:layout_margin="@dimen/activity_horizontal_margin"
-                android:id="@+id/totp"
-                android:textSize="24sp"
-                style="@style/OneTouchPrimaryTextStyle"
-                android:layout_width="match_parent"
-                android:layout_height="60dp"/>
-        <android.support.v4.view.ViewPager
-                android:id="@+id/container"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="1"/>
-    </LinearLayout>
-
-
-</android.support.design.widget.CoordinatorLayout>
