+++ twilio-auth-sample/src/main/res/menu/activity_main_drawer.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:showIn="navigation_view">
+
+    <group android:checkableBehavior="single">
+        <item
+            android:id="@+id/nav_requests"
+            android:icon="@drawable/ic_menu_requests"
+            android:title="@string/menu_navigation_requests" />
+        <item
+            android:id="@+id/nav_tokens"
+            android:icon="@drawable/ic_menu_tokens"
+            android:title="@string/menu_navigation_tokens" />
+    </group>
+</menu>
