--- twilio-auth-sample/src/main/res/menu/bottom_navigation_main_activity.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/menu_requests"
-        android:enabled="true"
-        android:icon="@drawable/default_logo_list_item"
-        android:title="@string/menu_bottom_navigation_requests"
-        app:showAsAction="ifRoom" />
-
-    <item
-        android:id="@+id/menu_tokens"
-        android:enabled="true"
-        android:icon="@drawable/default_logo_list_item"
-        android:title="@string/menu_bottom_navigation_tokens"
-        app:showAsAction="ifRoom" />
-</menu>
