+++ twilio-auth-sample/src/main/res/menu/menu_activity_main.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/menu_clear"
+        android:title="@string/menu_clear_data"
+        app:showAsAction="never" />
+</menu>
