+++ twilio-auth-sample/src/main/res/menu/menu_requests_fragment.xml
@@ -6,9 +6,4 @@
         android:icon="@drawable/ic_action_refresh"
         android:title="@string/menu_refresh_title"
         app:showAsAction="always" />
-
-    <item
-        android:id="@+id/menu_clear"
-        android:title="@string/menu_clear_data"
-        app:showAsAction="never" />
 </menu>
