+++ twilio-auth-sample/src/main/res/values/dimens.xml
@@ -2,13 +2,15 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
-    <dimen name="fab_margin">16dp</dimen>
     <dimen name="appbar_padding_top">8dp</dimen>
 
     <!-- Common values -->
     <dimen name="small_padding">4dp</dimen>
     <dimen name="large_padding">16dp</dimen>
     <dimen name="primary_action_text_size">16sp</dimen>
+
+    <!-- Main Activity -->
+    <dimen name="nav_header_height">176dp</dimen>
 
     <!-- Approval request list screen -->
     <dimen name="approval_request_item_height">88dp</dimen>
