+++ twilio-auth-sample/src/main/res/values/strings.xml
@@ -1,15 +1,16 @@
 <resources>
-    <string name="app_name">AuthySDKSample</string>
+    <string name="app_name">Twilio Auth SDK Sample</string>
 
     <string name="loading_msg">Loading…</string>
     <string name="error_no_connection">Please make sure you have internet connection</string>
 
     <!-- Main Activity -->
-    <string name="menu_bottom_navigation_tokens">Tokens</string>
-    <string name="menu_bottom_navigation_requests">Requests</string>
+    <string name="menu_navigation_tokens">Tokens</string>
+    <string name="menu_navigation_requests">Requests</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
 
-    <!-- Approval Requests List -->
-    <string name="title_activity_approval_requests_list">Approval Requests</string>
+    <!-- Requests List -->
     <string name="fragment_approval_requests_pending_title">PENDING</string>
     <string name="fragment_approval_requests_archive_title">ARCHIVE</string>
     <string name="transaction_logo_content_description">Transaction logo</string>
@@ -58,5 +59,4 @@
     <string name="registration_error_invalid_field">Make sure the value you entered is correct</string>
     <string name="registration_error_verification">Error registering the device</string>
     <string name="registration_error_device_deleted">Device deleted.\nEnter a new Authy ID and Backend URL</string>
-
 </resources>
